+++ CHANGES.md
@@ -1,5 +1,12 @@
 ### ChangeLog for GitHub Updater
 
+#### 2.4.0
+
+ * fixed transient assignment for tags returning empty array.
+ * added transient for `CHANGES.md` to themes, should further cut down on API 403 errors.
+ * new feature: theme rollback to previous version thanks @scarstens
+ * changed update methodology to use most recent tag first. If not tagged update from default branch.
+
 #### 2.3.3
 
  * fixed download link to have correct base URI for Repository Contents API. Oops.